Filter all expenses for only ones which are reimbursable
Filter list of all expenses for only ones which have been marked as
reimbursable. This was spurred by the fact that there was one duplicate
expense which was in a report but not marked as reimbursable. So it
would have not been reimbursed twice but still shows up in the list of
expenses. However, we still wanted it ignored. However, in general I
think this is what we want as any expenses not marked as reimbursable
would not actually be reimbursed and we would not want to tag them.